chore(create-muppet): minor improvements
